aya

Module maps

Data structures used to setup and share data with eBPF programs.

The eBPF platform provides data structures - maps in eBPF speak - that are used to setup and share data with eBPF programs. When you call Ebpf::load_file or Ebpf::load, all the maps defined in the eBPF code get initialized and can then be accessed using Ebpf::map, Ebpf::map_mut, or Ebpf::take_map.

Typed maps

The eBPF API includes many map types each supporting different operations. Ebpf::map, Ebpf::map_mut, and Ebpf::take_map always return the opaque &Map, &mut Map, and Map types respectively. Those three types can be converted to typed maps using the TryFrom or TryInto trait. For example:

use aya::maps::SockMap;
use aya::programs::SkMsg;

let intercept_egress = SockMap::try_from(bpf.map_mut("INTERCEPT_EGRESS").unwrap())?;
let map_fd = intercept_egress.fd().try_clone()?;
let prog: &mut SkMsg = bpf.program_mut("intercept_egress_packet").unwrap().try_into()?;
prog.load()?;
prog.attach(&map_fd)?;

Maps and Pod values

Many map operations copy data from kernel space to user space and vice versa. Because of that, all map values must be plain old data and therefore implement the Pod trait.

Re-exports

pub use array::Array;

pub use array::PerCpuArray;

pub use array::ProgramArray;

pub use bloom_filter::BloomFilter;

pub use hash_map::HashMap;

pub use hash_map::PerCpuHashMap;

pub use lpm_trie::LpmTrie;

pub use perf::AsyncPerfEventArray;

pub use perf::PerfEventArray;

pub use queue::Queue;

pub use ring_buf::RingBuf;

pub use sock::SockHash;

pub use sock::SockMap;

pub use stack::Stack;

pub use stack_trace::StackTraceMap;

pub use xdp::CpuMap;

pub use xdp::DevMap;

pub use xdp::DevMapHash;

pub use xdp::XskMap;

Modules

array

Array types.

bloom_filter

A Bloom Filter.

hash_map

Hash map types.

lpm_trie

A LPM Trie.

perf

Ring buffer types used to receive events from eBPF programs using the linux perf API.

queue

A FIFO queue.

ring_buf

A ring buffer map that may be used to receive events from eBPF programs. As of Linux 5.8, this is the preferred way to transfer per-event data from eBPF programs to userspace.

sock

Socket maps.

stack

A LIFO stack.

stack_trace

A hash map of kernel or user space stack traces.

xdp

XDP maps.

Structs

MapData

A generic handle to a BPF map.

MapFd

A map file descriptor.

MapInfo

Provides Provides metadata information about a loaded eBPF map.

MapIter

Iterator returned by map.iter().

MapKeys

Iterator returned by map.keys().

PerCpuValues

A slice of per-CPU values.

Enums

Map

eBPF map types.

MapError

Errors occuring from working with Maps

MapType

The type of eBPF map.

Traits

IterableMap

An iterable map

Functions

loaded_maps

Returns an iterator of MapInfo over all eBPF maps on the host.