aya::maps::array
Struct ProgramArray

pub struct ProgramArray<T> { /* private fields */ }

An array of eBPF program file descriptors used as a jump table.

eBPF programs can jump to other programs calling bpf_tail_call(ctx, prog_array, index). You can use ProgramArray to configure which programs correspond to which jump indexes.

Minimum kernel version

The minimum kernel version required to use this feature is 4.2.

Examples

use aya::maps::ProgramArray;
use aya::programs::CgroupSkb;

let mut prog_array = ProgramArray::try_from(bpf.take_map("JUMP_TABLE").unwrap())?;
let prog_0: &CgroupSkb = bpf.program("example_prog_0").unwrap().try_into()?;
let prog_0_fd =  prog_0.fd().unwrap();
let prog_1: &CgroupSkb = bpf.program("example_prog_1").unwrap().try_into()?;
let prog_1_fd = prog_1.fd().unwrap();
let prog_2: &CgroupSkb = bpf.program("example_prog_2").unwrap().try_into()?;
let prog_2_fd = prog_2.fd().unwrap();
let flags = 0;

// bpf_tail_call(ctx, JUMP_TABLE, 0) will jump to prog_0
prog_array.set(0, &prog_0_fd, flags);

// bpf_tail_call(ctx, JUMP_TABLE, 1) will jump to prog_1
prog_array.set(1, &prog_1_fd, flags);

// bpf_tail_call(ctx, JUMP_TABLE, 2) will jump to prog_2
prog_array.set(2, &prog_2_fd, flags);

Implementations

impl<T: Borrow<MapData>> ProgramArray<T>

pub fn indices(&self) -> MapKeys<'_, u32>

An iterator over the indices of the array that point to a program. The iterator item type is Result<u32, MapError>.

impl<T: BorrowMut<MapData>> ProgramArray<T>

pub fn set( &mut self, index: u32, program: &ProgramFd, flags: u64, ) -> Result<(), MapError>

Sets the target program file descriptor for the given index in the jump table.

When an eBPF program calls bpf_tail_call(ctx, prog_array, index), control flow will jump to program.

pub fn clear_index(&mut self, index: &u32) -> Result<(), MapError>

Clears the value at index in the jump table.

Calling bpf_tail_call(ctx, prog_array, index) on an index that has been cleared returns an error.

impl<T: Borrow<MapData>> ProgramArray<T>

pub fn pin<P: AsRef<Path>>(self, path: P) -> Result<(), PinError>

Pins the map to a BPF filesystem.

When a map is pinned it will remain loaded until the corresponding file is deleted. All parent directories in the given path must already exist.