aya::maps
Enum Map

pub enum Map {

    Array(MapData),
    BloomFilter(MapData),
    CpuMap(MapData),
    DevMap(MapData),
    DevMapHash(MapData),
    HashMap(MapData),
    LpmTrie(MapData),
    LruHashMap(MapData),
    PerCpuArray(MapData),
    PerCpuHashMap(MapData),
    PerCpuLruHashMap(MapData),
    PerfEventArray(MapData),
    ProgramArray(MapData),
    Queue(MapData),
    RingBuf(MapData),
    SockHash(MapData),
    SockMap(MapData),
    Stack(MapData),
    StackTraceMap(MapData),
    Unsupported(MapData),
    XskMap(MapData),

}

eBPF map types.

Variants

Array(MapData)

An Array map.

BloomFilter(MapData)

A BloomFilter map.

CpuMap(MapData)

A CpuMap map.

DevMap(MapData)

A DevMap map.

DevMapHash(MapData)

A DevMapHash map.

HashMap(MapData)

A HashMap map.

LpmTrie(MapData)

A LpmTrie map.

LruHashMap(MapData)

A HashMap map that uses a LRU eviction policy.

PerCpuArray(MapData)

A PerCpuArray map.

PerCpuHashMap(MapData)

A PerCpuHashMap map.

PerCpuLruHashMap(MapData)

A PerCpuHashMap map that uses a LRU eviction policy.

PerfEventArray(MapData)

A PerfEventArray map.

ProgramArray(MapData)

A ProgramArray map.

Queue(MapData)

A Queue map.

RingBuf(MapData)

A RingBuf map.

SockHash(MapData)

A SockHash map

SockMap(MapData)

A SockMap map.

Stack(MapData)

A Stack map.

StackTraceMap(MapData)

A StackTraceMap map.

Unsupported(MapData)

An unsupported map type.

XskMap(MapData)

A XskMap map.

Implementations

impl Map

pub fn pin<P: AsRef<Path>>(&self, path: P) -> Result<(), PinError>

Pins the map to a BPF filesystem.

When a map is pinned it will remain loaded until the corresponding file is deleted. All parent directories in the given path must already exist.