aya::maps
Function loaded_maps

pub fn loaded_maps() -> impl Iterator<Item = Result<MapInfo, MapError>>

Returns an iterator of MapInfo over all eBPF maps on the host.

Unlike Ebpf::maps, this includes all maps on the host system, not just those tied to a specific crate::Ebpf instance.

Uses kernel v4.13 features.

Example

for m in loaded_maps() {
    match m {
        Ok(map) => println!("{:?}", map.name_as_str()),
        Err(e) => println!("Error iterating maps: {:?}", e),
    }
}

Errors

Returns MapError::SyscallError if any of the syscalls required to either get next map id, get the map fd, or the MapInfo fail.

In cases where iteration can’t be performed, for example the caller does not have the necessary privileges, a single item will be yielded containing the error that occurred.