aya::maps::ring_buf
Struct RingBuf

pub struct RingBuf<T> { /* private fields */ }

A map that can be used to receive events from eBPF programs.

This is similar to crate::maps::PerfEventArray, but different in a few ways:

It’s shared across all CPUs, which allows a strong ordering between events.

Data notifications are delivered precisely instead of being sampled for every N events; the eBPF program can also control notification delivery if sampling is desired for performance reasons. By default, a notification will be sent if the consumer is caught up at the time of committing. The eBPF program can use the BPF_RB_NO_WAKEUP or BPF_RB_FORCE_WAKEUP flags to control this behavior.

On the eBPF side, it supports the reserve-commit pattern where the event can be directly written into the ring without copying from a temporary location.

Dropped sample notifications go to the eBPF program as the return value of reserve/output, and not the userspace reader. This might require extra code to handle, but allows for more flexible schemes to handle dropped samples.

To receive events you need to:

Construct RingBuf using RingBuf::try_from.

Call RingBuf::next to poll events from the RingBuf.

To receive async notifications of data availability, you may construct an tokio::io::unix::AsyncFd from the RingBuf’s file descriptor and poll it for readiness.

Minimum kernel version

The minimum kernel version required to use this feature is 5.8.

Examples

use aya::maps::RingBuf;
use std::convert::TryFrom;

let ring_buf = RingBuf::try_from(bpf.map_mut("ARRAY").unwrap()).unwrap();
let mut poll = poll_fd(ring_buf);
loop {
    let mut guard = poll.readable();
    let ring_buf = guard.inner_mut();
    while let Some(item) = ring_buf.next() {
        println!("Received: {:?}", item);
    }
    guard.clear_ready();
}

Polling

In the example above the implementations of poll(), poll.readable(), guard.inner_mut(), and guard.clear_ready() are not given. RingBuf implements the AsRawFd trait, so you can implement polling using any crate that can poll file descriptors, like epoll, mio etc. The above example API is motivated by that of tokio::io::unix::AsyncFd.

Implementations

impl<T> RingBuf<T>

pub fn next(&mut self) -> Option<RingBufItem<'_>>

Try to take a new entry from the ringbuf.

Returns Some(item) if the ringbuf is not empty. Returns None if the ringbuf is empty, in which case the caller may register for availability notifications through epoll or other APIs. Only one RingBufItem may be outstanding at a time.