aya::maps
Struct MapInfo

pub struct MapInfo(/* private fields */);

Provides Provides metadata information about a loaded eBPF map.

Introduced in kernel v4.13.

Implementations

impl MapInfo

pub fn from_id(id: u32) -> Result<Self, MapError>

Loads map info from a map ID.

Uses kernel v4.13 features.

pub fn map_type(&self) -> Result<MapType, MapError>

The type of map.

Introduced in kernel v4.13.

pub fn id(&self) -> u32

The unique ID for this map.

Introduced in kernel v4.13.

pub fn key_size(&self) -> u32

The key size for this map in bytes.

Introduced in kernel v4.13.

pub fn value_size(&self) -> u32

The value size for this map in bytes.

Introduced in kernel v4.13.

pub fn max_entries(&self) -> u32

The maximum number of entries in this map.

Introduced in kernel v4.13.

pub fn map_flags(&self) -> u32

The flags used in loading this map.

Introduced in kernel v4.13.

pub fn name(&self) -> &[u8]

The name of the map, limited to 16 bytes.

Introduced in kernel v4.15.

pub fn name_as_str(&self) -> Option<&str>

The name of the map as a &str.

None is returned if the name was not valid unicode or if field is not available.

Introduced in kernel v4.15.

pub fn fd(&self) -> Result<MapFd, MapError>

Returns a file descriptor referencing the map.

The returned file descriptor can be closed at any time and doing so does not influence the life cycle of the map.

Uses kernel v4.13 features.

pub fn from_pin<P: AsRef<Path>>(path: P) -> Result<Self, MapError>

Loads a map from a pinned path in bpffs.

Uses kernel v4.4 and v4.13 features.