aya::maps::xdp
Struct XskMap

pub struct XskMap<T> { /* private fields */ }

An array of AF_XDP sockets.

XDP programs can use this map to redirect packets to a target AF_XDP socket using the XDP_REDIRECT action.

Minimum kernel version

The minimum kernel version required to use this feature is 4.18.

Examples

use aya::maps::XskMap;

let mut xskmap = XskMap::try_from(bpf.map_mut("SOCKETS").unwrap())?;
// socket_fd is the RawFd of an AF_XDP socket
xskmap.set(0, socket_fd, 0);

See also

Kernel documentation: https://docs.kernel.org/next/bpf/map_xskmap.html

Implementations

impl<T: Borrow<MapData>> XskMap<T>

pub fn len(&self) -> u32

Returns the number of elements in the array.

This corresponds to the value of bpf_map_def::max_entries on the eBPF side.

impl<T: BorrowMut<MapData>> XskMap<T>

pub fn set( &mut self, index: u32, socket_fd: impl AsRawFd, flags: u64, ) -> Result<(), MapError>

Sets the AF_XDP socket at a given index.

When redirecting a packet, the AF_XDP socket at index will recieve the packet. Note that it will do so only if the socket is bound to the same queue the packet was recieved on.

Errors

Returns MapError::OutOfBounds if index is out of bounds, MapError::SyscallError if bpf_map_update_elem fails.

impl<T: Borrow<MapData>> XskMap<T>

pub fn pin<P: AsRef<Path>>(self, path: P) -> Result<(), PinError>

Pins the map to a BPF filesystem.

When a map is pinned it will remain loaded until the corresponding file is deleted. All parent directories in the given path must already exist.