aya

Module programs

eBPF program types.

eBPF programs are loaded inside the kernel and attached to one or more hook points. Whenever the hook points are reached, the programs are executed.

Loading and attaching programs

When you call Ebpf::load_file or Ebpf::load, all the programs included in the object code are parsed and relocated. Programs are not loaded automatically though, since often you will need to do some application specific setup before you can actually load them.

In order to load and attach a program, you need to retrieve it using Ebpf::program_mut, then call the load() and attach() methods, for example:

use aya::{Ebpf, programs::KProbe};

let mut bpf = Ebpf::load_file("ebpf_programs.o")?;
// intercept_wakeups is the name of the program we want to load
let program: &mut KProbe = bpf.program_mut("intercept_wakeups").unwrap().try_into()?;
program.load()?;
// intercept_wakeups will be called every time try_to_wake_up() is called
// inside the kernel
program.attach("try_to_wake_up", 0)?;

The signature of the attach() method varies depending on what kind of program you’re trying to attach.

Re-exports

pub use crate::programs::cgroup_device::CgroupDevice;

pub use crate::programs::cgroup_skb::CgroupSkb;

pub use crate::programs::cgroup_skb::CgroupSkbAttachType;

pub use crate::programs::cgroup_sock::CgroupSock;

pub use crate::programs::cgroup_sock_addr::CgroupSockAddr;

pub use crate::programs::cgroup_sockopt::CgroupSockopt;

pub use crate::programs::cgroup_sysctl::CgroupSysctl;

pub use crate::programs::extension::Extension;

pub use crate::programs::extension::ExtensionError;

pub use crate::programs::fentry::FEntry;

pub use crate::programs::fexit::FExit;

pub use crate::programs::kprobe::KProbe;

pub use crate::programs::kprobe::KProbeError;

pub use crate::programs::links::CgroupAttachMode;

pub use crate::programs::links::Link;

pub use crate::programs::links::LinkOrder;

pub use crate::programs::lirc_mode2::LircMode2;

pub use crate::programs::lsm::Lsm;

pub use crate::programs::perf_event::PerfEvent;

pub use crate::programs::perf_event::PerfEventScope;

pub use crate::programs::perf_event::PerfTypeId;

pub use crate::programs::perf_event::SamplePolicy;

pub use crate::programs::raw_trace_point::RawTracePoint;

pub use crate::programs::sk_lookup::SkLookup;

pub use crate::programs::sk_msg::SkMsg;

pub use crate::programs::sk_skb::SkSkb;

pub use crate::programs::sk_skb::SkSkbKind;

pub use crate::programs::sock_ops::SockOps;

pub use crate::programs::socket_filter::SocketFilter;

pub use crate::programs::socket_filter::SocketFilterError;

pub use crate::programs::tc::SchedClassifier;

pub use crate::programs::tc::TcAttachType;

pub use crate::programs::tc::TcError;

pub use crate::programs::tp_btf::BtfTracePoint;

pub use crate::programs::trace_point::TracePoint;

pub use crate::programs::trace_point::TracePointError;

pub use crate::programs::uprobe::UProbe;

pub use crate::programs::uprobe::UProbeError;

pub use crate::programs::xdp::Xdp;

pub use crate::programs::xdp::XdpError;

pub use crate::programs::xdp::XdpFlags;

Modules

cgroup_device

Cgroup device programs.

cgroup_skb

Cgroup skb programs.

cgroup_sock

Cgroup socket programs.

cgroup_sock_addr

Cgroup socket address programs.

cgroup_sockopt

Cgroup socket option programs.

cgroup_sysctl

Cgroup sysctl programs.

extension

Extension programs.

fentry

Fentry programs.

fexit

Fexit programs.

kprobe

Kernel space probes.

links

Program links.

lirc_mode2

Lirc programs.

lsm

LSM probes.

perf_attach

Perf attach links.

perf_event

Perf event programs.

raw_trace_point

Raw tracepoints.

sk_lookup

Programmable socket lookup.

sk_msg

Skmsg programs.

sk_skb

Skskb programs.

sock_ops

Socket option programs.

socket_filter

Socket filter programs.

tc

Network traffic control programs.

tp_btf

BTF-enabled raw tracepoints.

trace_point

Tracepoint programs.

uprobe

User space probes.

xdp

eXpress Data Path (XDP) programs.

Structs

ProgramFd

A Program file descriptor.

ProgramId

A Program identifier.

ProgramInfo

Provides metadata information about a loaded eBPF program.

Enums

CgroupSockAddrAttachType

Defines where to attach a CgroupSockAddr program.

CgroupSockAttachType

Defines where to attach a CgroupSock program.

CgroupSockoptAttachType

Defines where to attach a CgroupSockopt program.

ProbeKind

Kind of probe program

Program

eBPF program type.

ProgramError

Error type returned when working with programs.

ProgramType

The type of eBPF program.

Traits

MultiProgLink

Trait implemented by the Link types which support the kernel’s generic multi-prog API.

MultiProgram

Trait implemented by the Program types which support the kernel’s generic multi-prog API.

Functions

loaded_programs

Returns an iterator of ProgramInfo over all eBPF programs loaded on the host.