aya::programs
Function loaded_programs

pub fn loaded_programs() -> impl Iterator<Item = Result<ProgramInfo, ProgramError>>

Returns an iterator of ProgramInfo over all eBPF programs loaded on the host.

Unlike Ebpf::programs, this includes all programs on the host system, not just those tied to a specific crate::Ebpf instance.

Uses kernel v4.13 features.

Example

for p in loaded_programs() {
    match p {
        Ok(program) => println!("{}", String::from_utf8_lossy(program.name())),
        Err(e) => println!("Error iterating programs: {:?}", e),
    }
}

Errors

Returns ProgramError::SyscallError if any of the syscalls required to either get next program id, get the program fd, or the ProgramInfo fail.

In cases where iteration can’t be performed, for example the caller does not have the necessary privileges, a single item will be yielded containing the error that occurred.