aya::programs::links
Struct FdLink

pub struct FdLink { /* private fields */ }

A file descriptor link.

Fd links are returned directly when attaching some program types (for instance crate::programs::cgroup_skb::CgroupSkb), or can be obtained by converting other link types (see the TryFrom implementations).

An important property of fd links is that they can be pinned. Pinning can be used keep a link attached “in background” even after the program that has created the link terminates.

Example

 use aya::{Ebpf, programs::{links::FdLink, KProbe}};

 let program: &mut KProbe = bpf.program_mut("intercept_wakeups").unwrap().try_into()?;
 program.load()?;
 let link_id = program.attach("try_to_wake_up", 0)?;
 let link = program.take_link(link_id).unwrap();
 let fd_link: FdLink = link.try_into().unwrap();
 fd_link.pin("/sys/fs/bpf/intercept_wakeups_link").unwrap();

Implementations

impl FdLink

pub fn pin<P: AsRef<Path>>(self, path: P) -> Result<PinnedLink, PinError>

Pins the link to a BPF file system.

When a link is pinned it will remain attached even after the link instance is dropped, and will only be detached once the pinned file is removed. To unpin, see PinnedLink::unpin().

The parent directories in the provided path must already exist before calling this method, and must be on a BPF file system (bpffs).

Example

let link_id = prog.attach()?;
let owned_link = prog.take_link(link_id)?;
let fd_link: FdLink = owned_link.into();
let pinned_link = fd_link.pin("/sys/fs/bpf/example")?;