aya::programs::raw_trace_point
Struct RawTracePoint

pub struct RawTracePoint { /* private fields */ }

A program that can be attached at a pre-defined kernel trace point.

Unlike TracePoint, the kernel does not pre-process the arguments before calling the program.

The kernel provides a set of pre-defined trace points that eBPF programs can be attached to. See/sys/kernel/debug/tracing/events for a list of which events can be traced.

Minimum kernel version

The minimum kernel version required to use this feature is 4.17.

Examples

use aya::programs::RawTracePoint;

let program: &mut RawTracePoint = bpf.program_mut("sys_enter").unwrap().try_into()?;
program.load()?;
program.attach("sys_enter")?;

Implementations

impl RawTracePoint

pub fn load(&mut self) -> Result<(), ProgramError>

Loads the program inside the kernel.

pub fn attach( &mut self, tp_name: &str, ) -> Result<RawTracePointLinkId, ProgramError>

Attaches the program to the given tracepoint.

The returned value can be used to detach, see RawTracePoint::detach.

pub fn detach( &mut self, link_id: RawTracePointLinkId, ) -> Result<(), ProgramError>

Detaches from a tracepoint.

See RawTracePoint::attach.

pub fn take_link( &mut self, link_id: RawTracePointLinkId, ) -> Result<RawTracePointLink, ProgramError>

Takes ownership of the link referenced by the provided link_id.

The link will be detached on Drop and the caller is now responsible for managing its lifetime.

impl RawTracePoint

pub fn unload(&mut self) -> Result<(), ProgramError>

Unloads the program from the kernel.

Links will be detached before unloading the program. Note that owned links obtained using take_link() will not be detached.

impl RawTracePoint

pub fn fd(&self) -> Result<&ProgramFd, ProgramError>

Returns the file descriptor of this Program.

impl RawTracePoint

pub fn pin<P: AsRef<Path>>(&mut self, path: P) -> Result<(), PinError>

Pins the program to a BPF filesystem.

When a BPF object is pinned to a BPF filesystem it will remain loaded after Aya has unloaded the program. To remove the program, the file on the BPF filesystem must be removed. Any directories in the the path provided should have been created by the caller.

pub fn unpin(self) -> Result<(), Error>

Removes the pinned link from the filesystem.

impl RawTracePoint

pub fn from_pin<P: AsRef<Path>>(path: P) -> Result<Self, ProgramError>

Creates a program from a pinned entry on a bpffs.

Existing links will not be populated. To work with existing links you should use crate::programs::links::PinnedLink.

On drop, any managed links are detached and the program is unloaded. This will not result in the program being unloaded from the kernel if it is still pinned.

impl RawTracePoint

pub fn info(&self) -> Result<ProgramInfo, ProgramError>

Returns metadata information of this program.

Uses kernel v4.13 features.