aya::programs::tc
Function qdisc_detach_program

pub fn qdisc_detach_program(
    if_name: &str,
    attach_type: TcAttachType,
    name: &str,
) -> Result<(), Error>

Detaches the programs with the given name.

Errors

Returns io::ErrorKind::NotFound to indicate that no programs with the given name were found, so nothing was detached. Other error kinds indicate an actual failure while detaching a program.