aya::programs

Trait MultiProgLink

pub trait MultiProgLink {
    // Required method
    fn fd(&self) -> Result<BorrowedFd<'_>, LinkError>;
}

Trait implemented by the Link types which support the kernel’s generic multi-prog API.

Minimum kernel version

The minimum kernel version required to use this feature is 6.6.0.

Required Methods

fn fd(&self) -> Result<BorrowedFd<'_>, LinkError>

Borrows the file descriptor.