aya::util
Function syscall_prefix

pub fn syscall_prefix() -> Result<&'static str, Error>

👎Deprecated since 0.12.0: On some systems - commonly on 64 bit kernels that support running 32 bit applications - the syscall prefix depends on what architecture an application is compiled for, therefore attaching to only one prefix is incorrect and can lead to security issues.

Returns the prefix used by syscalls.

Example

use aya::util::syscall_prefix;
let prefix = syscall_prefix().unwrap();
let syscall_fname = format!("{prefix}exec");

Errors

Returns std::io::ErrorKind::NotFound if the prefix can’t be guessed. Returns other std::io::Error kinds if /proc/kallsyms can’t be opened or is somehow invalid.