Crate aya

A library to work with eBPF programs.

eBPF is a technology that allows running user-supplied programs inside the Linux kernel. For more info see https://ebpf.io/what-is-ebpf.

Aya is an eBPF library built with a focus on operability and developer experience. It does not rely on libbpf nor bcc - it’s built from the ground up purely in Rust, using only the libc crate to execute syscalls. With BTF support and when linked with musl, it offers a true compile once, run everywhere solution, where a single self-contained binary can be deployed on many linux distributions and kernel versions.

Some of the major features provided include:

Support for the BPF Type Format (BTF), which is transparently enabled when supported by the target kernel. This allows eBPF programs compiled against one kernel version to run on different kernel versions without the need to recompile.

Support for function call relocation and global data maps, which allows eBPF programs to make function calls and use global variables and initializers.

Async support with both tokio and async-std.

Easy to deploy and fast to build: aya doesn’t require a kernel build or compiled headers, and not even a C toolchain; a release build completes in a matter of seconds.

Modules

maps

Data structures used to setup and share data with eBPF programs.

pin

Pinning BPF objects to the BPF filesystem.

programs

eBPF program types.

sys

A collection of system calls for performing eBPF related operations.

util

Utility functions.

Macros

include_bytes_aligned

Include bytes from a file for use in a subsequent crate::Ebpf::load.

Structs

Btf

BPF Type Format metadata.

Ebpf

The main entry point into the library, used to work with eBPF programs and maps.

EbpfLoader

Builder style API for advanced loading of eBPF programs.

GlobalData

Global data that can be exported to eBPF programs before they are loaded.

VerifierLogLevel

Used to set the verifier log level flags in EbpfLoader.

bpf_map_def

Map definition in legacy BPF map declaration style

Enums

BtfError

The error type returned when BTF operations fail.

EbpfError

The error type returned by Ebpf::load_file and Ebpf::load.

Endianness

An endianness that is selectable at run-time.

PinningType

The pinning type

Traits

Pod

Marker trait for types that can safely be converted to and from byte slices.

Functions

features

Returns a reference to the detected BPF features.

Type Aliases

BpfDeprecated

The main entry point into the library, used to work with eBPF programs and maps.

BpfErrorDeprecated

The error type returned by Bpf::load_file and Bpf::load.

BpfLoaderDeprecated

Builder style API for advanced loading of eBPF programs.