aya::maps::array
Struct Array

pub struct Array<T, V: Pod> { /* private fields */ }

A fixed-size array.

The size of the array is defined on the eBPF side using the bpf_map_def::max_entries field. All the entries are zero-initialized when the map is created.

Minimum kernel version

The minimum kernel version required to use this feature is 3.19.

Examples

use aya::maps::Array;

let mut array = Array::try_from(bpf.map_mut("ARRAY").unwrap())?;
array.set(1, 42, 0)?;
assert_eq!(array.get(&1, 0)?, 42);

Implementations

impl<T: Borrow<MapData>, V: Pod> Array<T, V>

pub fn len(&self) -> u32

Returns the number of elements in the array.

This corresponds to the value of bpf_map_def::max_entries on the eBPF side.

pub fn get(&self, index: &u32, flags: u64) -> Result<V, MapError>

Returns the value stored at the given index.

Errors

Returns MapError::OutOfBounds if index is out of bounds, MapError::SyscallError if bpf_map_lookup_elem fails.

pub fn iter(&self) -> impl Iterator<Item = Result<V, MapError>> + '_

An iterator over the elements of the array. The iterator item type is Result<V, MapError>.

impl<T: BorrowMut<MapData>, V: Pod> Array<T, V>

pub fn set( &mut self, index: u32, value: impl Borrow<V>, flags: u64, ) -> Result<(), MapError>

Sets the value of the element at the given index.

Errors

Returns MapError::OutOfBounds if index is out of bounds, MapError::SyscallError if bpf_map_update_elem fails.

impl<T: Borrow<MapData>, V: Pod> Array<T, V>

pub fn pin<P: AsRef<Path>>(self, path: P) -> Result<(), PinError>

Pins the map to a BPF filesystem.

When a map is pinned it will remain loaded until the corresponding file is deleted. All parent directories in the given path must already exist.