aya::sys
Function enable_stats

pub fn enable_stats(stats_type: Stats) -> Result<OwnedFd, SyscallError>

Enable global statistics tracking for eBPF programs and returns a file descriptor handler.

Statistics tracking is disabled when the file descriptor is dropped (either automatically when the variable goes out of scope or manually through Drop).

Usage:

Obtain fd from enable_stats and bind it to a variable.

Record the statistic of interest.

Wait for a recorded period of time.

Record the statistic of interest again, and calculate the difference.

Close/release fd automatically or manually.

Introduced in kernel v5.8.

Examples

use aya::sys::{enable_stats, Stats};

let _fd = enable_stats(Stats::RunTime)?;